+++ TLM/TLM/State/Flags.cs
@@ -315,7 +315,7 @@ namespace TrafficManager.State {
 #if DEBUGFLAGS
             Log._Debug($"Flags.resetLaneArrowFlags: Resetting lane arrows of lane {laneId}.");
 #endif
-            if (LaneConnectionManager.Instance.HasOutgoingConnections(laneId)) {
+            if (LaneConnectionManager.Instance.Sub.HasOutgoingConnections(laneId)) {
                 return false;
             }
 
@@ -392,7 +392,7 @@ namespace TrafficManager.State {
                 return false; // disallow custom lane arrows in highway rule mode
             }
 
-            if (LaneConnectionManager.Instance.HasOutgoingConnections(laneId, startNode)) {
+            if (LaneConnectionManager.Instance.Sub.HasOutgoingConnections(laneId, startNode)) {
                 // TODO refactor
                 res = SetLaneArrow_Result.LaneConnection;
                 return false; // custom lane connection present
