+++ TLM/TLM/UI/SubTools/LaneArrows/LaneArrowTool.cs
@@ -269,7 +269,7 @@ namespace TrafficManager.UI.SubTools.LaneArrows {
                 sort: false);
 
             foreach (var lane in lanes) {
-                if (!LaneConnectionManager.Instance.HasOutgoingConnections(lane.laneId))
+                if (!LaneConnectionManager.Instance.Sub.HasOutgoingConnections(lane.laneId))
                     return true;
             }
