+++ TLM/TLM/Util/Record/LaneConnectionRecord.cs
@@ -19,7 +19,7 @@ namespace TrafficManager.Util.Record {
 
         private static LaneConnectionManager connMan => LaneConnectionManager.Instance;
 
-        private uint[] GetCurrentConnections() => connMan.GetLaneConnections(LaneId, StartNode);
+        private uint[] GetCurrentConnections() => connMan.Sub.GetLaneConnections(LaneId, StartNode);
 
         public void Record() {
             connections_ = GetCurrentConnections();
@@ -31,7 +31,7 @@ namespace TrafficManager.Util.Record {
 
         public void Restore() {
             if (connections_ == null) {
-                connMan.RemoveLaneConnections(LaneId, StartNode);
+                connMan.Sub.RemoveLaneConnections(LaneId, StartNode);
                 return;
             }
             var currentConnections = GetCurrentConnections();
@@ -40,12 +40,12 @@ namespace TrafficManager.Util.Record {
 
             foreach (uint targetLaneId in connections_) {
                 if (currentConnections == null || !currentConnections.Contains(targetLaneId)) {
-                    connMan.AddLaneConnection(LaneId, targetLaneId, StartNode);
+                    connMan.Sub.AddLaneConnection(LaneId, targetLaneId, StartNode);
                 }
             }
             foreach (uint targetLaneId in currentConnections ?? Enumerable.Empty<uint>()) {
                 if (!connections_.Contains(targetLaneId)) {
-                    connMan.RemoveLaneConnection(LaneId, targetLaneId, StartNode);
+                    connMan.Sub.RemoveLaneConnection(LaneId, targetLaneId, StartNode);
                 }
             }
         }
@@ -61,7 +61,7 @@ namespace TrafficManager.Util.Record {
             var mappedLaneId = MappedLaneId(LaneId);
 
             if (connections_ == null) {
-                connMan.RemoveLaneConnections(mappedLaneId, StartNode);
+                connMan.Sub.RemoveLaneConnections(mappedLaneId, StartNode);
                 return;
             }
 
@@ -74,7 +74,7 @@ namespace TrafficManager.Util.Record {
                 if (mappedTargetLaneId == 0)
                     continue;
                 //Log._Debug($"connecting lanes: {mappedLaneId}->{mappedTargetLaneId}");
-                connMan.AddLaneConnection(mappedLaneId, mappedTargetLaneId, StartNode);
+                connMan.Sub.AddLaneConnection(mappedLaneId, mappedTargetLaneId, StartNode);
             }
         }
